StorageTestSuite -> FullTestSuite
